"PRODIGY_TrackCode_TaskNumber
This program converts temperatures between Celsius, Fahrenheit, and Kelvin scales. 
The user is prompted to input a temperature value and the original unit of measurement. 
The program then converts the temperature to the other two units and displays the converted values to the user.

For example, if the user enters a temperature of 25 degrees Celsius, the program converts it to Fahrenheit and Kelvin, 
and presents the converted values as outputs.